+++ moor/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Changed the sql representation of text types from `VARCHAR` to `TEXT`
 - Added extensions for `isNull` and `isNotNull`
+- Support creating a `VmDatabase` from a raw sqlite3 `Database` via `VmDatabase.opened`
 
 ## 3.4.0